# Conflicts:
#   core/pom.xml
#   core/src/main/java/fr/uge/db/insert/monitoring/MonitorInserter.java

+++ core/src/main/java/fr/uge/modules/api/server/external/model/ReportParameter.java
@@ -1,4 +1,5 @@
 package fr.uge.modules.api.server.external.model;
 
-public record ReportParameter(boolean expanded, boolean cache, int proximity_limit, int network_size) {
+public record ReportParameter(boolean expanded, boolean cache, float proximity_limit, int network_size) {
+
 }
